+++ CHANGELOG.md
@@ -5,9 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- Implementation for AIS VDM/VDO sentence type 20 parsing
-- Implementation for AIS VDM/VDO sentence type 22 parsing
+- Implementation for AIS VDM/VDO sentence type 25 parsing
 - Implementation for AIS VDM/VDO sentence type 23 parsing
+- Implementation for AIS VDM/VDO sentence type 22 parsing
+- Implementation for AIS VDM/VDO sentence type 20 parsing
 
 ## [0.6.0] - 2020-11-18
 ### Added